* Components/Searching/NSSet+ArticleFiltering.m         * Components/ArticleTable/ArticleTablePlugin.m           some NSInteger conversions, and clang warning shutup
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2223 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Grr/Components/Searching/NSSet+ArticleFiltering.m
@@ -33,9 +33,9 @@
 - (NSSet*) subsetFilteredForString: (NSString*) searchString
 {
 	NSArray* articles;
-	unsigned subsetSize = 0;
+	NSUInteger subsetSize = 0;
 	id<Article>* subsetObjects;
-	int i;
+	NSUInteger i;
 	NSSet* resultSet;
 
 	if (searchString == nil || [searchString length] == 0) {
@@ -57,7 +57,7 @@
 	        }
 	    }
 	}
-	NSLog(@"filtered down to %d of %d articles.", subsetSize, [articles count]);
+	NSLog(@"filtered down to %d of %" PRIuPTR " articles.", subsetSize, [articles count]);
 	resultSet = [NSSet setWithObjects: subsetObjects count: subsetSize];
 
 	free( subsetObjects );
